Add "noautoparent" restore command option to disable the automatic parent directory selection
At this time, when we select a directory, we mark parent directories for
the restore automatically.

In the NDMP restore process, at this time, when we select a directory
entry, the entire content is restored recursively.

This option can be used to not select automatically parent directories
during a restore session.