Implementaion of .ls command for Plugins.
The plugin=... parameter is added to .ls command which should be the name of
the plugin with all required plugin parameters in the form found in fileset
Plugin=... directive. The plugin should implement a special listing=<path>
parameter for estimate job to use this functionality. The .ls command path=...
parameter is sent as listing=<path> for Plugin, so user can inform a Plugin
what he expecits to listing.