Add the RestoreClient directive for Restore job.
With this directive you can define a default restore client
used with restore job. If this directive is not set then a
default restore client will be set to a backup client as usual.
Now it is possible to define a dedicated restore job and run
an automatic (scheduled) restore tests of your backups which
will be redirected to restore tests system.