Fix #3593 VirtualFull will select jobs to consolidate using Job name in addition to Client/FileSet
If multiple jobs on a client are sharing the same fileset, the virtual full
was selecting all the jobs matching the Client/FileSet names. Leading to
a VirtualFull with the correct files, but an incorrect parent hierarchy.

The new code will select the JobIds to consolidate with the job Name,
the Client and the FileSet.