Add command to change the pool of a job and the associated volumes
update job jobid=x pool=XXXX

It will update the PoolId for the Job and all associated volumes.

This might be very useful to do VirtualFull in a temporary pool,
then move back the job and all the volumes into a selected pool.

The main use case is when using VirtualFull + MaximumPoolBytes