Fix #3615 about bconsole Socket errors reported in the bacula log file
The idea is to send the messages into a specific place to keep track
of them. Unfortunately, customers are opening a lot of tickets to know
if the messages are important or not, and how to solve the problem.

Waiting for a final implementation, socket errors for consoles
will be sent to the trace file instead of the log file.