Refactoring of BSOCK and introducing BSOCKCORE.
A BSOCKCORE class was designed as a code split from previous BSOCK class
and as a base class for new BSOCK. The new BSOCK class should retain
virtually the same API as the old one. This was achieved (not)as a
simple class variables and methods split between two new classes. All
the methods and code which implement Bacula communication protocol goes
to the new BSOCK. This includes the following functionalities:
communication line compression, data spooling and authentication. All
the rest as a generic functions and methods goes to BSOCKCORE.