Fix verify volume jobs with sparse files
The verify read() loop was not reading the data like the backup loop was
doing. So the sparse block detection was miss-aligned giving false errors.

We now use the configuration NetworkBufferSize to adjust the amount of data we
read in a loop. I'm not sure it's 100% accurate for dedup jobs.