Add Docker Plugin for FileDaemon.
This is a File Daemon Docker Plugin which backup and restore
Docker objects including images, containers and persistent
volumes data with dedicated docker image. The plugin requires
a docker utility for compile and build. The default is to
automatically check for docker binary and if found then build
the plugin. Additionally refactor some configuration and build
procedures.