Fix copy/migration job selection
Description:
-----------

The Copy/Migration job selection with a Job name was selecting
Jobs that are already migrated. It is leading to a situation where
a new catalog record is created for the new job, but the job is not
used at the end. JobFiles=0 JobBytes=0.

In some conditions, the accurate code selects the incorrect job
record.

In the catalog, both Job records have the same attributes
(same JobTDate, same StartTime). So the selection process
was returning one or the other depending on the weather cast.

We no longer try to copy/migrate a job that is not a valid
Backup job, and we check that the Job status is T.