Fix db_get_accurate_jobids() with concurrent queries on the same jobs
The btemp3 table name is computed with the from_jobid argument in
the .bvfs_get_jobid command. If we have multiple requests at the
same time, as the tables are not TEMPORARY (due to a MySQL issue),
one query can drop the table of the other one and lead to problems.

Waiting to have the from_jobid feature fully implemented, the code
is removed.